Added user control
Move around a little red ball

+++ src/ca/mkmxwl/zeppelin/entities/Entity.java
@@ -4,7 +4,7 @@ import java.awt.*;
 
 public class Entity {
 
-	double x, y, dir;
+	int x, y, dir;
 
 	public void update() {
 	}
